organized files, added coverage, added linting, stubbed tests
