+++ __tests__/AddSomeTests.test.js
@@ -1,0 +1,4 @@
+
+test('We have tests', () => {
+    expect(false).toBeTruthy();
+});
