+++ index.js
@@ -1,8 +1,8 @@
-import Container from './Container';
+import Container from './src/Container';
 
 const NAMESPACE = "translationNotes";
 
 export default {
   name: NAMESPACE,
   container: Container
-}
+};
