+++ package.json
@@ -11,13 +11,41 @@
   "main": "index.js",
   "display": "app",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "eslint ./src ./index.js && jest"
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx}",
+      "!**/node_modules/**"
+    ],
+    "coverageDirectory": "coverage"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/translationCoreApps/translationNotes_Check_plugin.git"
+  },
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/translationCoreApps/translationNotes_Check_plugin/issues"
+  },
+  "homepage": "https://github.com/translationCoreApps/translationNotes_Check_plugin#readme",
+  "devDependencies": {
+    "babel-eslint": "^8.0.0",
+    "babel-jest": "^21.0.2",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "eslint": "^4.7.2",
+    "eslint-plugin-react": "^7.4.0",
+    "jest": "^21.1.0",
+    "react-test-renderer": "^15.6.1"
   },
   "dependencies": {
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.16.0",
     "natural": "^0.4.0",
-    "react": "^15.4.2",
+    "react": "^15.6.1",
     "regexp": "^1.0.0",
     "xregexp": "^3.1.1"
   }
