+++ src/scripts/UpdateSectionList.js
